fixed error to ensure that inactive player does not keep their hand into the next round if their partner goes alone. starting to play with holdem logic.
